Add files via upload
The core of the dashboard lies in the key metrics required for monitoring. Thus, based on 
whether the dashboard is for an organization on the whole or for a department such as sales, 
finance, human resources, production, etc. the key metrics that are required for display vary. 
Further, the key metrics for a dashboard also depend on the role of the recipients (audience). 
For example, Executive (CEO, CIO, etc.), Operations Manager, Sales Head, Sales Manager, 
etc. This is due to the fact that the primary goal of a dashboard in to enable data visualization 
for decision making. 
The success of a dashboard often depends on the metrics that were chosen for monitoring. For 
example, Key Performance Indicators, Balanced Scorecards and Sales Performance Figures 
could be the content appropriate in business dashboards.